feat: refactor admin components, services, and models into an "admin" directory
