+++ src/app/admin/admin.module.ts
@@ -1,0 +1,32 @@
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
+import { SharedModule } from './../shared/shared.module';
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { AdminProductsComponent } from './components/admin-products/admin-products.component';
+import { AdminOrdersComponent } from './components/admin-orders/admin-orders.component';
+import { ProductFormComponent } from './components/product-form/product-form.component';
+import { AdminAuthGuard } from './services/admin-auth-guard.service';
+import { MatPaginatorModule } from '@angular/material/paginator';
+import { MatTableModule } from '@angular/material/table';
+import { MatSortModule } from '@angular/material/sort';
+import { AppRoutingModule } from 'app/app-routing.module';
+
+@NgModule({
+  declarations: [
+    AdminProductsComponent,
+    AdminOrdersComponent,
+    ProductFormComponent,
+  ],
+  imports: [
+    CommonModule,
+    SharedModule,
+    FormsModule,
+    ReactiveFormsModule,
+    MatPaginatorModule,
+    MatTableModule,
+    MatSortModule,
+    AppRoutingModule,
+  ],
+  providers: [AdminAuthGuard],
+})
+export class AdminModule {}
