+++ src/app/admin/components/admin-products/admin-products.component.scss
