+++ src/app/app-routing.module.ts
@@ -1,6 +1,6 @@
-import { AdminAuthGuard } from './admin-auth-guard.service';
+import { AdminAuthGuard } from './admin/services/admin-auth-guard.service';
 import { MyOrdersComponent } from './my-orders/my-orders.component';
-import { AdminProductsComponent } from './admin/admin-products/admin-products.component';
+import { AdminProductsComponent } from './admin/components/admin-products/admin-products.component';
 import { LoginComponent } from './login/login.component';
 import { OrderSuccessComponent } from './order-success/order-success.component';
 import { ShoppingCartComponent } from './shopping-cart/shopping-cart.component';
@@ -9,9 +9,9 @@ import { HomeComponent } from './home/home.component';
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 import { CheckOutComponent } from './check-out/check-out.component';
-import { AdminOrdersComponent } from './admin/admin-orders/admin-orders.component';
+import { AdminOrdersComponent } from './admin/components/admin-orders/admin-orders.component';
 import { AuthGuard } from 'shared/services/auth-guard.service';
-import { ProductFormComponent } from './admin/product-form/product-form.component';
+import { ProductFormComponent } from './admin/components/product-form/product-form.component';
 
 const routes: Routes = [
   { path: '', component: ProductsComponent },
