+++ src/app/app.module.ts
@@ -1,58 +1,49 @@
-import { MyOrdersComponent } from './my-orders/my-orders.component';
-import { AdminOrdersComponent } from './admin/admin-orders/admin-orders.component';
-import { OrderService } from 'shared/services/order.service';
-import { CheckOutComponent } from './check-out/check-out.component';
-import { ShoppingCartService } from 'shared/services/shopping-cart.service';
-import { ShoppingCartComponent } from './shopping-cart/shopping-cart.component';
-import { ProductsComponent } from './products/products.component';
-import { ProductService } from 'shared/services/product.service';
-import { CategoryService } from 'shared/services/category.service';
-import { AdminProductsComponent } from './admin/admin-products/admin-products.component';
-import { UserService } from 'shared/services/user.service';
-import { AuthGuard } from 'shared/services/auth-guard.service';
-import { AuthService } from 'shared/services/auth.service';
-import { environment } from './../environments/environment';
-import { BrowserModule } from '@angular/platform-browser';
+import { AdminModule } from './admin/admin.module';
+import { SharedModule } from './shared/shared.module';
 import { NgModule } from '@angular/core';
 import { AngularFireModule } from '@angular/fire';
-import { AngularFireDatabaseModule } from '@angular/fire/database';
 import { AngularFireAuthModule } from '@angular/fire/auth';
+import { AngularFireDatabaseModule } from '@angular/fire/database';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
+import { MatPaginatorModule } from '@angular/material/paginator';
+import { MatSortModule } from '@angular/material/sort';
+import { MatTableModule } from '@angular/material/table';
+import { BrowserModule } from '@angular/platform-browser';
+import { NoopAnimationsModule } from '@angular/platform-browser/animations';
+import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
+
+import { environment } from './../environments/environment';
+import { AdminAuthGuard } from './admin/services/admin-auth-guard.service';
+import { AdminOrdersComponent } from './admin/components/admin-orders/admin-orders.component';
+import { AdminProductsComponent } from './admin/components/admin-products/admin-products.component';
+import { ProductFormComponent } from './admin/components/product-form/product-form.component';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { BsNavbarComponent } from './bs-navbar/bs-navbar.component';
-import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
-import { AdminAuthGuard } from './admin-auth-guard.service';
-import { ProductFormComponent } from './admin/product-form/product-form.component';
-import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { NoopAnimationsModule } from '@angular/platform-browser/animations';
-import { MatPaginatorModule } from '@angular/material/paginator';
-import { MatTableModule } from '@angular/material/table';
-import { MatSortModule } from '@angular/material/sort';
+import { CheckOutComponent } from './check-out/check-out.component';
+import { MyOrdersComponent } from './my-orders/my-orders.component';
 import { ProductFilterComponent } from './products/product-filter/product-filter.component';
-import { ProductCardComponent } from 'shared/components/product-card/product-card.component';
-import { ProductQuantityComponent } from 'shared/components/product-quantity/product-quantity.component';
-import { ShoppingCartSummaryComponent } from './shopping-cart-summary/shopping-cart-summary.component';
+import { ProductsComponent } from './products/products.component';
 import { ShippingFormComponent } from './shipping-form/shipping-form.component';
+import { ShoppingCartSummaryComponent } from './shopping-cart-summary/shopping-cart-summary.component';
+import { ShoppingCartComponent } from './shopping-cart/shopping-cart.component';
 
 @NgModule({
   declarations: [
     AppComponent,
     BsNavbarComponent,
-    AdminProductsComponent,
     CheckOutComponent,
-    ProductFormComponent,
     ProductsComponent,
     ProductFilterComponent,
-    ProductCardComponent,
     ShoppingCartComponent,
-    ProductQuantityComponent,
     ShoppingCartSummaryComponent,
     ShippingFormComponent,
-    AdminOrdersComponent,
     MyOrdersComponent,
   ],
   imports: [
     BrowserModule,
+    SharedModule,
+    AdminModule,
     FormsModule,
     ReactiveFormsModule,
     AppRoutingModule,
@@ -61,20 +52,8 @@ import { ShippingFormComponent } from './shipping-form/shipping-form.component';
     AngularFireAuthModule,
     NgbModule,
     NoopAnimationsModule,
-    MatPaginatorModule,
-    MatTableModule,
-    MatSortModule,
   ],
-  providers: [
-    AuthService,
-    AuthGuard,
-    UserService,
-    AdminAuthGuard,
-    CategoryService,
-    ProductService,
-    ShoppingCartService,
-    OrderService,
-  ],
+  providers: [],
   bootstrap: [AppComponent],
 })
 export class AppModule {}
